pyproject.toml
[project]
name = "study-django"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "qe-present", email = "2664481691@qq.com" }
]
dependencies = [
    "django>=5.1.6",
    "pyinstaller>=6.12.0",
    "django-cors-headers>=4.7.0",
    "django-rest-framework>=0.1.0",
]
readme = "README.md"
requires-python = ">= 3.13"

[tool.rye.scripts]

dev = "python manage.py runserver"
build="pyinstaller -F manage.py"
build_add_data='pyinstaller -F --add-data="./templates;templates" --add-data="./static;static" manage.py'


